feat: secure delete_todo with ownership verification
Refactor by Adedara Benson. Added user_dependency to the delete endpoint. Updated database query to filter by both todo_id and owner_id to prevent unauthorized data deletion.

+++ routers/todos.py
@@ -114,10 +114,15 @@ async def read_todo_by_priority(db: db_dependency,
 We can decide to delete some data from the data base by todo_id
 """
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency ,todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency,db: db_dependency ,todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=404, detail='Authentication failed ')
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
